APIGeneradorDeCuestionarios/requirements.txt
absl-py==0.10.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
astroid==2.2.5
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.8.2
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.1
coverage==5.0.3
cycler==0.10.0
cymem==2.0.4
DateTime==4.3
decorator==4.4.2
distlib==0.3.1
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.1
dnspython==1.16.0
docutils==0.16
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-2.3.1/es_core_news_lg-2.3.1.tar.gz
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-2.3.1/es_core_news_md-2.3.1.tar.gz
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.3.1/es_core_news_sm-2.3.1.tar.gz
filelock==3.0.12
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.17.2
joblib==0.16.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.2.2
matplotlib==3.1.3
mccabe==0.6.1
murmurhash==1.0.4
mysql-connector-python==8.0.21
nltk==3.5
numpy==1.18.1
pandas==1.1.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.3
prompt-toolkit==3.0.7
protobuf==3.6.1
pycodestyle==2.6.0
pygame==1.9.6
Pygments==2.7.1
pylint==2.3.1
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMuPDF==1.16.11
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.13.0
PyQt5-sip==12.8.0
pytesseract==0.3.2
python-dateutil==2.8.1
pythonnet==2.4.0
pytz==2020.1
pywin32==228
pyzmq==19.0.2
regex==2020.10.28
requests==2.22.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.12.0
soupsieve==1.9.5
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.3
statistics==1.0.3.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
toga==0.3.0.dev18
toga-core==0.3.0.dev18
toga-demo==0.3.0.dev19
toga-winforms==0.3.0.dev18
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.4
travertino==0.1.2
typed-ast==1.3.4
urllib3==1.25.8
virtualenv==20.0.30
Wand==0.5.9
wasabi==0.8.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.11.1
zipp==3.1.0
zope.interface==5.1.0
